requirements.txt
datasets==2.3.2
numpy==1.19.5
pandas==1.1.5
regex==2021.4.4
scikit_learn
torch==1.11.0
transformers==4.20.1
tez
AttrDict
bitsandbytes-cuda110
sentencepiece
protobuf==3.20.0
psutil
tqdm
AttrDict
bitsandbytes-cuda113
